Reduce the scope of `QueryRewriteContext`
Today we provide a lot of functinality on the QueryRewirteContext that
we potentially don't have ie. if we rewrite on a coordinating node or when
we percolating. This change moves most of the unnecessary shard level or
index level services and dependencies to `QueryShardContext` instead.
