+++ core/src/main/java/org/elasticsearch/index/query/QueryShardContext.java
@@ -50,6 +50,7 @@ import org.elasticsearch.index.query.support.NestedScope;
 import org.elasticsearch.index.similarity.SimilarityService;
 import org.elasticsearch.script.Script;
 import org.elasticsearch.script.ScriptService;
+import org.elasticsearch.script.TemplateScript;
 import org.elasticsearch.search.lookup.SearchLookup;
 
 import java.io.IOException;
@@ -66,12 +67,14 @@ import static java.util.Collections.unmodifiableMap;
  */
 public class QueryShardContext extends QueryRewriteContext {
 
+    private final ScriptService scriptService;
+    private final IndexSettings indexSettings;
     private final MapperService mapperService;
     private final SimilarityService similarityService;
     private final BitsetFilterCache bitsetFilterCache;
     private final IndexFieldDataService indexFieldDataService;
-    private final IndexSettings indexSettings;
     private final int shardId;
+    private final IndexReader reader;
     private String[] types = Strings.EMPTY_ARRAY;
     private boolean cachable = true;
     private final SetOnce<Boolean> frozen = new SetOnce<>();
@@ -94,15 +97,17 @@ public class QueryShardContext extends QueryRewriteContext {
             IndexFieldDataService indexFieldDataService, MapperService mapperService, SimilarityService similarityService,
             ScriptService scriptService, NamedXContentRegistry xContentRegistry,
             Client client, IndexReader reader, LongSupplier nowInMillis) {
-        super(indexSettings, mapperService, scriptService, xContentRegistry, client, reader, nowInMillis);
+        super(xContentRegistry, client, nowInMillis);
         this.shardId = shardId;
-        this.indexSettings = indexSettings;
         this.similarityService = similarityService;
         this.mapperService = mapperService;
         this.bitsetFilterCache = bitsetFilterCache;
         this.indexFieldDataService = indexFieldDataService;
         this.allowUnmappedFields = indexSettings.isDefaultAllowUnmappedFields();
         this.nestedScope = new NestedScope();
+        this.scriptService = scriptService;
+        this.indexSettings = indexSettings;
+        this.reader = reader;
 
     }
 
@@ -315,7 +320,6 @@ public class QueryShardContext extends QueryRewriteContext {
     }
 
     /** Return the script service to allow compiling scripts. */
-    @Override
     public final ScriptService getScriptService() {
         failIfFrozen();
         return scriptService;
@@ -347,10 +351,10 @@ public class QueryShardContext extends QueryRewriteContext {
         }
     }
 
-    @Override
     public final String getTemplateBytes(Script template) {
         failIfFrozen();
-        return super.getTemplateBytes(template);
+        TemplateScript compiledTemplate = scriptService.compile(template, TemplateScript.CONTEXT).newInstance(template.getParams());
+        return compiledTemplate.execute();
     }
 
     /**
@@ -382,4 +386,30 @@ public class QueryShardContext extends QueryRewriteContext {
     public QueryBuilder parseInnerQueryBuilder(XContentParser parser) throws IOException {
         return AbstractQueryBuilder.parseInnerQueryBuilder(parser);
     }
+
+    @Override
+    public final QueryShardContext convertToShardContext() {
+        return this;
+    }
+
+    /**
+     * Returns the index settings for this context. This might return null if the
+     * context has not index scope.
+     */
+    public IndexSettings getIndexSettings() {
+        return indexSettings;
+    }
+
+    /**
+     * Return the MapperService.
+     */
+    public MapperService getMapperService() {
+        return mapperService;
+    }
+
+    /** Return the current {@link IndexReader} */
+    public IndexReader getIndexReader() {
+        return reader;
+    }
+
 }
